Fix client vs server rendering issues in welcome page
