+++ apps/scoutgame/hooks/useIsMounted.tsx
@@ -1,0 +1,13 @@
+import { useCallback, useEffect, useRef, useState } from 'react';
+
+export function useIsMounted(): boolean {
+  const [isMounted, setIsMounted] = useState(false);
+
+  useEffect(() => {
+    if (!isMounted) {
+      setIsMounted(true);
+    }
+  }, [isMounted]);
+
+  return isMounted;
+}
